Define new structure and separate js from images and css
sources will be structured as it follows:
src
├── index.js // entrypoint for the Web Application
├── images
├── sass
└── js
    ├── App.jsx             // Root node for the App
    ├── Routes.jsx          // App's Routes
    ├── PrivateRoute.jsx
    ├── components
    │   ├── charts
    │   ├── pipeline
    │   ├── layout          // big layout pieces
    │   └── ui              // small ui elements (buttons, titles, badges...)
    ├── pages
    │   ├── templates       // templates to build App pages
    │   │   └── Page.jsx
    │   ├── auth
    |   |   └
    │   ├── Home.jsx
    │   └── Waiting.jsx
    └── services
        ├── api
        ├── colors.js
        ├── history.js
        ├── react-auth0-spa.js
        └── serviceWorker.js

Main changes:
- `js` is separated from `sass` and `images`
Other changes:
- `utils`, `serviceWorker` and `Auth0Provider` were moved into `services`
- `ui`  will contain small and repeatable elements (buttons, titles, badges...)
