+++ src/index.js
@@ -1,0 +1,5 @@
+import 'bootstrap/dist/js/bootstrap.js';
+
+import './sass/main.scss';
+
+require('./js/App.jsx');
