+++ src/js/App.jsx
@@ -1,15 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import 'bootstrap/dist/js/bootstrap.js';
+import Routes from './Routes';
 
-import './sass/main.scss';
-import App from './App';
-
-import * as serviceWorker from './serviceWorker';
-
-import { Auth0Provider } from './react-auth0-spa';
-import history from './utils/history';
+import * as serviceWorker from './services/serviceWorker';
+import { Auth0Provider } from './services/react-auth0-spa';
+import history from './services/history';
 
 // A function that routes the user to the right place
 // after login
@@ -28,7 +24,7 @@ ReactDOM.render(
     redirect_uri={window.ENV.auth.redirectUri}
     onRedirectCallback={onRedirectCallback}
   >
-    <App />
+    <Routes />
   </Auth0Provider>,
   document.getElementById('root')
 );
