+++ src/js/components/charts/CleanAreaChart.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { FlexibleXYPlot, AreaSeries, LineSeries, GradientDefs } from 'react-vis';
 
-import { hexToRGBA } from '../../utils/colors';
+import { hexToRGBA } from '../../services/colors';
 
 export default ({ fill, stroke, data }) => {
