+++ src/js/components/charts/TimeSeries.jsx
