+++ src/js/components/pipeline/Pipeline.jsx
