+++ src/js/pages/Waiting.jsx
@@ -1,8 +1,6 @@
 import React, { useEffect } from 'react';
 
-import 'bootstrap/dist/js/bootstrap.js';
-
-import Page from '../components/layout/Page';
+import Page from '../templates/Page';
 
 import { useHistory } from 'react-router-dom';
 
@@ -38,13 +36,13 @@ export default () => {
           <div id="carousel-waiting" className="carousel slide w-75 mx-auto" data-ride="carousel">
             <ol className="carousel-indicators">
               {slides.map((slide, index) => {
-                return <li key={index} data-target="#carousel-waiting" data-slide-to={index} className={index === 0 ? "active": ""}></li>;
+                return <li key={index} data-target="#carousel-waiting" data-slide-to={index} className={index === 0 ? "active" : ""}></li>;
               })}
             </ol>
             <div className="carousel-inner">
               {slides.map((slide, index) => {
                 return (
-                  <div key={index} className={index === 0 ? "carousel-item active": "carousel-item"}>
+                  <div key={index} className={index === 0 ? "carousel-item active" : "carousel-item"}>
                     <img className="d-block w-100" src={slide.src} alt={slide.alt} />
                   </div>
                 );
