+++ src/js/pages/auth/Logout.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
-import { useAuth0 } from '../../react-auth0-spa';
+import { useAuth0 } from '../../services/react-auth0-spa';
 
 export default () => {
-  const { loading, isAuthenticated } = useAuth0();
+  const { loading, isAuthenticated, logout } = useAuth0();
 
   if (loading) {
     return <div>Loading...</div>;
@@ -13,5 +13,6 @@ export default () => {
     return <div>Not authenticated!</div>;
   }
 
+  logout();
   return [];
 };
