+++ src/js/pages/templates/Page.jsx
@@ -1,9 +1,9 @@
 import React, { useState, useEffect } from 'react';
 
-import Navbar from './Navbar';
-import Footer from './Footer';
+import Navbar from '../components/layout/Navbar';
+import Footer from '../components/layout/Footer';
 
-import { getUser } from '../../services/api';
+import { getUser } from '../services/api';
 
 export default ({ children }) => {
   const [userState, setUser] = useState(null);
