+++ src/js/services/api/openapi-client/.openapi-generator/VERSION
