+++ src/js/services/api/openapi-client/model/CalculatedMetric.js
