+++ src/js/services/api/openapi-client/model/CalculatedMetricValues.js
