+++ src/js/services/api/openapi-client/model/CalculatedMetrics.js
