+++ src/js/services/api/openapi-client/model/Granularity.js
