+++ src/js/services/api/openapi-client/model/InvalidRequestErrorAllOf.js
